feat(mcp-utils): enhance type definitions and refactor imports
This commit introduces the McpServerConnection and McpStatus types to improve the structure and clarity of the MCP connection management. It also updates the relevant imports across the codebase to reflect these changes, ensuring consistency and better organization.
