+++ packages/common/src/mcp-utils/types.ts
@@ -6,9 +6,34 @@ import type {
   McpServerTransportStdio,
 } from "../configuration/index.js";
 
+export interface McpServerConnection {
+  status: "stopped" | "starting" | "ready" | "error";
+  error: string | undefined;
+  tools: {
+    [toolName: string]: McpToolStatus;
+  };
+}
+
 export interface McpToolStatus extends McpTool {
   disabled: boolean;
 }
+
+export type McpStatus = {
+  /**
+   * Connection status for each MCP server.
+   */
+  connections: {
+    [serverName: string]: McpServerConnection;
+  };
+  /**
+   * Reduced available toolset from all MCP servers, disabled tools are excluded.
+   */
+  toolset: {
+    [toolName: string]: McpTool;
+  };
+
+  instructions: string;
+};
 
 export function isStdioTransport(
   config: McpServerTransport,
