+++ packages/common/src/vscode-webui-bridge/index.ts
@@ -7,11 +7,6 @@ export { createVscodeHostStub } from "./webview-stub";
 export type { PochiCredentials } from "./types/pochi";
 
 export type {
-  McpStatus,
-  McpConnection,
-  McpToolStatus,
-} from "./types/mcp";
-export type {
   UserEditsDiff,
   ExecuteCommandResult,
   SaveCheckpointOptions,
