+++ packages/common/src/vscode-webui-bridge/webview.ts
@@ -5,7 +5,6 @@ import type { UserInfo } from "../configuration";
 import type {
   CaptureEvent,
   CustomAgentFile,
-  McpStatus,
   NewTaskParams,
   ResourceURI,
   RuleFile,
@@ -15,6 +14,7 @@ import type {
   UserEditsDiff,
   WorkspaceState,
 } from "./index";
+import type { McpStatus } from "../mcp-utils";
 import type { DisplayModel } from "./types/model";
 import type { PochiCredentials } from "./types/pochi";
